Updated geom_input to read and use user set conversion factors for au2ang and ang2au.
The default value is set to the inverse of 0.529177249 in machine
precision on cagle which is 1.88972598858d0

Also added to geom.F are geom_print_distances etc. angle dihedral
These are not called anywhere yet but testing will be completed
and then the functionality will be exported to stepper and driver.

rak


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@2662 85dd5168-0e29-45a6-bdf9-2122c44e486f