HvD: With the direct MP2 capability the code would always store the same amount of data no matter how many processors were being used. This is undesirable because for every pass all 2-electron integrals need to be recalculated.
The memory calculations are fixed up now so that when more processors
are available a larger fraction of the transformed 2-electron integrals
is stored. This can lead to a superlinear speedup with increasing
numbers of processors. In calculating the memory requirements one has
to be careful because the data distribution imposed insists that a whole
column or row be kept on a single processor. The memory limits have to
allow for this, rather than assume an even distribution.

Calculations on C60 showed that the 3rd index transformation in the
"twofold" (the current default) algorithm scales very poorly. The
"sixfold" algorithm scales much better. This makes me wonder whether
the current default is the optimal choice.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@19759 85dd5168-0e29-45a6-bdf9-2122c44e486f