HvD: Dummy centers vs Ghost centers. This fix changes the behavior of dummy centers and banquo centers. Originally dummy centers (named "x") were merely a mathematical trick to specify zmatrices for tricky molecules. E.g. benzene, or linear molecules. Ghost centers (named "bq" after the ghost Banquo from William Shakespeare's Macbeth) were introduced to calculate BSSE free energies. These centers have basis sets but no charge, in that sense they are not atoms but do change the mathematical representation of the problem.
Because of their role it is alright to delete dummy centers. This was introduced
at some point to clean up zmatrices. After deleting the dummies the autoz 
function would be used to regenerate a zmatrix. Unfortunately, the way the 
deletion was implemented meant that depending on how dummy centers were named
different things would happen. Due to a different choice of internal coordinates
this could actually lead to a different optimization path. 

To fix this all centers starting with "x" but not "xe" are deleted and a new
zmatrix generated. Bq centers by their nature should not be deleted, and they
now won't be. Not clear to me at present is what happens if the original input
zmatrix contains constraints. I don't see how these can be handled so that they
would always survive the elimination of the dummy centers. Of course one could
replace the dummy centers by ghost centers without a basis set to deal with 
such a case. The dft_bsse test case has been updated to match the new approach.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@19761 85dd5168-0e29-45a6-bdf9-2122c44e486f