HvD: In response to the development of a new ARMCI over MPI implementation I have updated the build_nwchem script and the tools GNUmakefile to be able to drive this target. This implementation goes by the name MPI_TS (short for MPI Two-Sided) as it eliminates the data-server and uses MPI two-sided communications to implement the ARMCI functionality. It is expected to be highly portable as it relies only on MPI. Trying the performance of this implementation will be interesting.
In all cases it is currently required to set the environment variable

    EXP_GA

any value for this variable will do. This variable is currently required because
the implementation lives in a separate branch of GA. In addition you need to
run get-tools with this environment variable set to pick up the correct GA
branch.

The build_nwchem script will under certain conditions automatically pick the
MPI_TS implementation up. However, it is safer to set 

    ARMCI_NETWORK=MPI_TS

as that eliminates any guess work that might lead to a different result. 

If you want to try this implementation, please go ahead, as I am sure the GA
team will value any feedback we provide.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@22632 85dd5168-0e29-45a6-bdf9-2122c44e486f