HvD: Turned the COSMO surface charge correction off by default.
In COSMO the total surface charge should be the same as the charge contained
within the surface. In practice this is not always obtained exactly and
previously a correction was applied to adjust the total surface charge. However,
as pointed out by Michel Dupuis, this correction is geometry dependent and not
taken into account when calculating gradients. The result is that geometry
optimizations often do not converge (neither analytically or numerically).
When the surface charge corrections are suppressed the geometry optimizations
are essentially trouble free. 

So these corrections seem a bad idea. Therefore the defaults are adjusted to
ensure the best ride when running the code with the default settings.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@22777 85dd5168-0e29-45a6-bdf9-2122c44e486f