HvD: Added the do_gasphase option to the COSMO capability. This option controls whether a COSMO calculation starts with the electronic structure of the gas phase molecule, or whether it jumps right in at the solvation stage.
In many instances it is OK to jump in at the solvation stage, but in some cases
fitting COSMO charges to an unconverged electron density leads to an environment
that locks the calculation into something unphysical. This happened for example
with a geometry optimization of H3CONa. Therefore we let the user control this.

If the gas phase calculation can be skipped in a geometry optimization it can 
speed the calculation up by a factor of 2 easily.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@22820 85dd5168-0e29-45a6-bdf9-2122c44e486f