HvD: Finally fixed the bug that caused the geometry optimizer to fail in case of symmetry and constrained geometric parameters. The problem was that if a symmetric starting structure was provided but parameters (bond lengths, bond angles, or torsion angles) were declared constant so that a geometry optimization would lower the point group the code failed. In those cases the code would enforce the point group symmetry of the input structure and ignore the constraints rather than adjust the point group symmetry to be compatible with the constraints.
The changes implemented make the symmetry code aware of any constraints imposed
on the geometry. This is done by generating a numeric label for each atom that
is different for each type of and number of constraints that an atom appears in.
For two atoms to be symmetry equivalent we insist that they must:
- be at positions that are related by the symmetry operators
- have the same tag
- have the same numeric constraint label.
The 4 calculations in the test case should all behave in the same way and give
the same result.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26303 85dd5168-0e29-45a6-bdf9-2122c44e486f