HvD: The compilation of nwxc_eval_print.F was causing problems. This file is needed only for the nwxc_fnc_print program that in turn is needed to extract the Maxima expressions for a density functional from the corresponding automatic differentiation Fortran implementation.
The issue with compiling this file lies in the usage of the nwad1p module.
The *.mod files that are being generated by compiling a Fortran module cause
that a single compiler invokation might generate multiple targets. I haven't
worked out how to best keep track of that in the Makefile. This caused 
nwxc_eval_print.F to be compiled before the nwad1p.mod file was generated 
which in turn caused the compilation to fail. Rather than putting a dependency
in I have decided that someone who wants to use it should manually run

	. ../../nwchem_make_env.sh
	make nwxc_fcn_print

This is a reasonable stance given that nwxc_fnc_print is only used under the
special circumstances when someone wants to generate a symbolic algebra 
generated version of a functional that is already implemented using the
automatic differentiation approach. This is not a routine use case of our code!


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26524 85dd5168-0e29-45a6-bdf9-2122c44e486f